Silence a warning in clang-3.5 dealing with implicit conversion from 64bit to 32bit in a function call to switch_ivr_sleep()

+++ src/switch_channel.c
@@ -3732,7 +3732,7 @@ SWITCH_DECLARE(switch_status_t) switch_channel_perform_answer(switch_channel_t *
 		const char *delay;
 
 		if ((delay = switch_channel_get_variable(channel, "answer_delay"))) {
-			long msec = atol(delay);
+			uint32_t msec = atoi(delay);
 			
 			if (msec) {
 				switch_ivr_sleep(channel->session, msec, SWITCH_TRUE, NULL);
